+++ lib/components/animation_component.dart
@@ -5,7 +5,7 @@ import 'package:flame/animation.dart';
 
 class AnimationComponent extends PositionComponent {
   Animation animation;
-  Paint paint;
+  Paint overridePaint;
   bool destroyOnFinish = false;
 
   AnimationComponent(double width, double height, this.animation,
@@ -54,9 +54,8 @@ class AnimationComponent extends PositionComponent {
   @override
   void render(Canvas canvas) {
     prepareCanvas(canvas);
-    animation
-        .getSprite()
-        .render(canvas, width: width, height: height, overridePaint: paint);
+    animation.getSprite().render(canvas,
+        width: width, height: height, overridePaint: overridePaint);
   }
 
   @override